Use a `ThreadLocal` shared list for absolute sequence propagation
Working under the following requirements:

- No changes to API
- Assumption that this is going to be a hot path in most projects (not just osu!) and should be expected to not allocate more than necessary
- Should work from load thread or update thread, as it currently does

This seems like about the best we can do. I can understand if this approach is a tad hard to swallow, but wanted to push this out for discussion. Using "static" lists with the paradigm I've used here is safe enough, but the lists do still have the ability to grow and reside in memory indefinitely. Maybe this is fine?

If we are not happy with `ThreadLocal`, it is also enough to cater to the most common usage (ie. only use a shared list on the `Update` thread, and construct a new one in all other cases).

Note that this is a similar issues to #4683 and we may want to decide on a path forward to use in cases like these as they come up. #4683 is more simple as it can only ever run on the update thread (and also know the length from the outset).