TODO: `ThreadRunner` with non-input main thread
Just some preliminary thinking, but if we are using a delegate callback
to run the draw work, there's a very high chance (as long as we are
running above our target frame rate in single thread mode) that we want
to use this callback to trigger *all* work, including input, update and
audio.

This is a bit of a larger change so I'm not going to do it immediately,
but it shouldn't be out of reach as we can replace the `InputThread`
provided to `ThreadRunner` with the `DrawThread` in such cases.