Attempt delegation local to `DrawThread`
This works well on macOS, but on iOS there's issues in both threading
modes where the input (main) thread slows to a grind. Still
investigating why.

That said, in single thread mode this is not going to hold. We likely
want to direct *all* game load to the draw delegate in such cases, so I
think the call to set the delegate will need to exist in `GameHost`
instead...